+++ common/stdafx.h
@@ -115,6 +115,7 @@
 // <algorithm>
 // <limits>
 // <string>
+// <string_view>
 // <type_traits>
 // <cstring>